requirements.txt
streamlit==1.24.0
pandas==1.5.3
mysql-connector-python==8.0.33
Pillow==9.5.0
python-dotenv==1.0.0
requests==2.29.0
boto3==1.28.0
botocore==1.31.0
